Fix #16982 - do not try to delete internal relations if they are not configured

+++ libraries/classes/Database/Designer/Common.php
@@ -773,6 +773,14 @@ class Common
             }
         }
 
+        // internal (pmadb) relation is not working, skip delete
+        if ($GLOBALS['cfgRelation']['relwork'] == false) {
+            return [
+                false,
+                __('Error: Relational features are disabled!'),
+            ];
+        }
+
         // internal relations
         $delete_query = 'DELETE FROM '
             . Util::backquote($GLOBALS['cfgRelation']['db']) . '.'
